Refactoring 5: Inline Method
Normally you would look for 'Extract Method' opportunities, but given
that we will refactor this more, it makes sense to Inline the private method
into the public method. Then we can refactor it more as we go.
NOTE: you could wait on this Inline Method refactoring until later, keeping the
other refactorings happening in the private method.