Refactoring 16: Convert Procedural Design to Objects
The public interface is still the class method generate() but internally
it now builds an instance of Primes to hold the array of primes and size variables.
Then the internal method calls don't need to pass these values.
NOTE: sieve() needs to be made public so that the class method can access it.